Migrating to Gradle and Docker.
