+++ .gitignore
@@ -1,21 +1,8 @@
-# Gradle
+# Ignore Gradle project-specific cache directory
 .gradle
-build/
 
-# Ignore Gradle GUI config
-gradle-app.setting
+# Ignore Gradle build output directory
+build
 
-# Eclipse
-/.classpath
-/.settings/
-/.project
-/bin/
-
-# IntelliJ
-.idea/workspace.xml
-*.ipr
-*.iws
-out
-
-# Misc
-*.log
+*.iml
+.idea
