--- .idea/libraries/lib.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="lib">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/lib/mvc-framework-1.0-SNAPSHOT.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/mysql-connector-java-8.0.21.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/mvc-framework-1.0-SNAPSHOT-javadoc.jar!/" />
-      <root url="jar://$PROJECT_DIR$/lib/mvc-framework-1.0-SNAPSHOT-sources.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
