+++ docker-compose.yaml
@@ -1,0 +1,10 @@
+version: "3"
+services:
+  db:
+    image: mysql:5.7.31
+    volumes:
+      - ./mysql-data:/var/lib/mysql
+    ports:
+      - "3306:3306"
+    environment:
+      - MYSQL_ROOT_PASSWORD=password
