+++ src/main/java/edu/cscc/views/home/HomeGoodbyeView.java
